+++ src/main/java/insane96mcp/iguanatweaksexpanded/module/experience/enchanting/ISEEnchantingTable.java
@@ -110,15 +110,11 @@ public class ISEEnchantingTable extends BaseEntityBlock {
                                 player.sendSystemMessage(Component.translatable("iguanatweaksexpanded.enchanting_table.already_known", enchantmentDescId.setStyle(Style.EMPTY.withColor(ChatFormatting.DARK_PURPLE).withHoverEvent(new HoverEvent(HoverEvent.Action.SHOW_TEXT, Component.translatable("iguanatweaksexpanded.enchanting_table.currently", lvlKnownDescId))))));
                             continue;
                         }
-                        //if (lvlKnown == lvl)
-                        //    lvl++;
                         MutableComponent newLvlDescId = Component.translatable("enchantment.level." + lvl).withStyle(ChatFormatting.DARK_PURPLE);
                         if (EnchantingFeature.isConsumedOnEnchant(enchantment)) {
                             int currentLvl = enchantingTableBE.knownEnchantments.get(enchantment);
-                            MutableComponent currentLvlDescId = Component.translatable("enchantment.level." + currentLvl).withStyle(ChatFormatting.DARK_PURPLE);
-                            if (currentLvl > 10) {
-                                currentLvlDescId = Component.literal(currentLvl + "").withStyle(Style.EMPTY.withColor(ChatFormatting.DARK_PURPLE));
-                            }
+                            newLvlDescId = Component.literal(lvl + "").withStyle(Style.EMPTY.withColor(ChatFormatting.DARK_PURPLE));
+                            MutableComponent currentLvlDescId = Component.literal(currentLvl + "").withStyle(Style.EMPTY.withColor(ChatFormatting.DARK_PURPLE));
                             player.sendSystemMessage(Component.translatable("iguanatweaksexpanded.enchanting_table.added_levels", newLvlDescId, enchantmentDescId, currentLvlDescId));
                         }
                         else if (lvlKnown == 0) {